Breakpad processor: Use unsigned constants in comparisons, to quiet compiler warnings.
This patch avoids comparisons between signed and unsigned values, as
warned about by G++ 4.4.1.

a=jimblandy, r=nealsid


git-svn-id: http://google-breakpad.googlecode.com/svn/trunk@469 4c0a9323-5329-0410-9bdc-e9ce6186880e

+++ src/processor/minidump_unittest.cc
@@ -65,7 +65,7 @@ TEST_F(MinidumpTest, TestMinidumpFromFile) {
   ASSERT_TRUE(minidump.Read());
   const MDRawHeader* header = minidump.header();
   ASSERT_NE(header, (MDRawHeader*)NULL);
-  ASSERT_EQ(header->signature, MD_HEADER_SIGNATURE);
+  ASSERT_EQ(header->signature, u_int32_t(MD_HEADER_SIGNATURE));
   //TODO: add more checks here
 }
 
@@ -91,7 +91,7 @@ TEST_F(MinidumpTest, TestMinidumpFromStream) {
   ASSERT_TRUE(minidump.Read());
   const MDRawHeader* header = minidump.header();
   ASSERT_NE(header, (MDRawHeader*)NULL);
-  ASSERT_EQ(header->signature, MD_HEADER_SIGNATURE);
+  ASSERT_EQ(header->signature, u_int32_t(MD_HEADER_SIGNATURE));
   //TODO: add more checks here
 }
